requirements.txt
easydict==1.9
fastdtw==0.3.4
matplotlib==3.5.2
numpy==1.21.2
pandas==1.4.3
PyYAML==6.0
scikit-learn==0.24.0
scipy==1.8.1
torch==1.11.0
tqdm==4.60.0